use pgbouncer to pool connections
pgbouncer is a separate service that can pool our connections to the
database, this'll help prevent queuepool errors when we're under load.

we connect to pgbouncer like a regular database, but we have to manage
the credentials separately since it's an app on our paas rather than
a vcap service like the current rds.

we use nullpool to tell sqlalchemy to not manage any connections, and
just open one connection per each worker thread.