Change to the _status endpoint.
The _status endpoint returns data about the deployed build which includes the dB version. The dB is queried to get the alembic_version, this is a quick query but is a query none the less.
I'm not convinced that this endpoint needs to use a dB connection to return this data.
I also renamed the attributes of the json to stop referring to travis.

Sound like a good idea?