Use the original version of boto
We aren't aware of any reason we need to use our fork of boto anymore.
We therefore swap to use `celery[sqs]` which brings in the original
version of boto and we can remove our use of the fork.

This has been tested by running the celery app and seeing it connect to
sqs and grab messages off a queue.