The reporting app populates the stats table each evening.
The `create-nightly-notification-status` task, kicks off 4 tasks for email for 4 days plus 4 tasks for sms plus 10 tasks for letters for a total of 18 tasks. Since there are 2 deivery-worker-reporting apps with 11 worker threads all 18tasks are running in parallel. This put unnecessary pressure on the app, recently we had the app crash with a `out of memory` error. If we reduce the number of worker threads this will all the tasks to be run 4 at a time.