set preload_app=True in gunicorn config
preload_app is a flag that makes the gunicorn master process load the
app code, and then copy that code to each worker. This is opposed to the
default, where each worker initialises the app itself. We hope that by
enabling this flag, we'll solve issues we've seen where instances
sometimes return 502s when they start up. Our gut feel is that this
happens because the workers are listening on the socket before they've
finished loading the config etc. If the master thread loads the config,
there shouldn't be this issue.

Advantages of preload_app: "save some RAM resources as well as speed up
server boot times" [1]

Disadvantages of preload_app. Can no longer live-reload gunicorn
workers to get new code. We don't do this anyway.

[1] https://docs.gunicorn.org/en/stable/settings.html#preload-app