re-run failed reporting tasks by marking in redis when they last ran
We've had issues with long running tasks dying for a variety of reasons,
and us having to manually kick things off again.

when a task runs, we store a key in redis. The key looks like
`task_last_run_timestamp_{task_name}_{kwargs}` where kwargs are
`key_value` separeated by underscores. The value is the timestamp of
when the task completed.

Then, every morning at 6am BST we can run a task that checks the
timestamps of all the individual day tasks. If that timestamp is older
than midnight, the task didn't run this morning, so kick it off again.
To kick it off again it just runs the parent tasks with a new flag
"only_run_missing_days", to keep code duplication down.

The reason that each task and param gets an individual key is so that
each one can be given an expiry date. If it was a hash set inside redis,
then we wouldn't have a nice automated way of expiring individual keys
without doing the work ourselves.

I've set the expiry date to four days, so that we can check on things
after the weekend if everything goes really wrong.