Don’t interface with serialised data directly
This commit copies the `JSONModel` class from the admin app, which turns
serialised data (a dict made from JSON) into an object on which certain
predefined properties are allowed.

This means we can still do the caching of serialised data, without
having to change too much of the code in the app, or make it ugly by
sprinkling dict lookups everywhere.