Add gds-pre-commit file to repo
https://github.com/alphagov/gds-pre-commit

This should help us avoid accidently checking in any real secrets. It
has detected secrets in other repos at GDS so worth us trialling.

As part of this, it is suggested that we check in the
`.secrets.baseline` file as this will mean that the state of which
strings are safe is stored in the repo and won't be flagged as non safe
by any other team mates checking.